Miniter/Miniter_WAS/requirements.txt
amqp==5.0.9
arrow==1.2.1
attrs==21.4.0
Automat==20.2.0
awscli==1.22.26
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
botocore==1.23.26
celery==5.2.3
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
constantly==15.1.0
docutils==0.15.2
Flask @ file:///home/ktietz/src/ci/flask_1611932660458/work
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
Flask-Twisted==0.1.2
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
importlib-metadata @ file:///opt/concourse/worker/volumes/live/b1c84e32-1519-4554-50d3-980dc7c220d9/volume/importlib-metadata_1631916711680/work
incremental==21.3.0
iniconfig==1.1.1
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1624781299557/work
jmespath==0.10.0
kombu==5.2.3
MarkupSafe @ file:///opt/concourse/worker/volumes/live/1bcf0940-5bf8-4171-61e3-2133f5885e8b/volume/markupsafe_1621528148241/work
mysql-connector-python==8.0.27
observable==1.0.3
packaging==21.3
pika==1.2.0
pluggy==1.0.0
prompt-toolkit==3.0.24
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==1.4.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.29
stravalib==0.10.4
toml==0.10.2
Twisted==21.7.0
typing_extensions==4.0.1
units==0.7
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
Werkzeug @ file:///tmp/build/80754af9/werkzeug_1631818832573/work
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
zope.interface==5.4.0
